+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Next Release
 
+### Breaking Changes
+- Remove --validator-ids option from watermark-repair subcommand [#909](https://github.com/Consensys/web3signer/pull/909)
+
 ### Features Added
 - Aws bulk loading for secp256k1 keys in eth1 mode [#889](https://github.com/Consensys/web3signer/pull/889)